Beginnings of the shuffle
