+++ workspace/TS100/Core/Inc/FRToSI2C.hpp
@@ -25,10 +25,12 @@ public:
 
 	static void CpltCallback(); //Normal Tx Callback
 
-	static void Mem_Read(uint16_t DevAddress, uint16_t MemAddress,
+	static bool Mem_Read(uint16_t DevAddress, uint16_t MemAddress,
 			uint16_t MemAddSize, uint8_t *pData, uint16_t Size);
 	static void Mem_Write(uint16_t DevAddress, uint16_t MemAddress,
 			uint16_t MemAddSize, uint8_t *pData, uint16_t Size);
+	//Returns true if device ACK's being addressed
+	static bool probe(uint16_t DevAddress);
 
 	static void Transmit(uint16_t DevAddress, uint8_t *pData, uint16_t Size);
 	static void I2C_RegisterWrite(uint8_t address, uint8_t reg, uint8_t data);
