+++ workspace/TS100/Core/Inc/MMA8652FC.hpp
@@ -10,7 +10,7 @@
 #include "stm32f1xx_hal.h"
 #include "MMA8652FC_defines.h"
 #include "FRToSI2C.hpp"
-#include "hardware.h"
+#include "BSP.h"
 
 class MMA8652FC {
