+++ workspace/TS100/Core/Inc/gui.hpp
@@ -9,12 +9,11 @@
 #define GUI_HPP_
 #include "Translation.h"
 #include "Settings.h"
-#include "hardware.h"
+#include "BSP.h"
 
 #define PRESS_ACCEL_STEP			3
 #define PRESS_ACCEL_INTERVAL_MIN	10
 #define PRESS_ACCEL_INTERVAL_MAX	30
-
 
 //GUI holds the menu structure and all its methods for the menu itself
 
@@ -33,6 +32,7 @@ typedef struct {
 } menuitem;
 
 void enterSettingsMenu();
+void GUIDelay();
 extern const menuitem rootSettingsMenu[];
 
 #endif /* GUI_HPP_ */
