+++ workspace/TS100/Core/Src/FRToSI2C.cpp
@@ -4,10 +4,10 @@
  *  Created on: 14Apr.,2018
  *      Author: Ralim
  */
-#include "hardware.h"
+#include "BSP.h"
 #include "FRToSI2C.hpp"
 #define I2CUSESDMA
-I2C_HandleTypeDef* FRToSI2C::i2c;
+I2C_HandleTypeDef *FRToSI2C::i2c;
 SemaphoreHandle_t FRToSI2C::I2CSemaphore;
 StaticSemaphore_t FRToSI2C::xSemaphoreBuffer;
 
@@ -18,13 +18,14 @@ void FRToSI2C::CpltCallback() {
 	}
 }
 
-void FRToSI2C::Mem_Read(uint16_t DevAddress, uint16_t MemAddress,
-		uint16_t MemAddSize, uint8_t* pData, uint16_t Size) {
+bool FRToSI2C::Mem_Read(uint16_t DevAddress, uint16_t MemAddress,
+		uint16_t MemAddSize, uint8_t *pData, uint16_t Size) {
 
 	if (I2CSemaphore == NULL) {
 		// no RToS, run blocking code
 		HAL_I2C_Mem_Read(i2c, DevAddress, MemAddress, MemAddSize, pData, Size,
 				5000);
+		return true;
 	} else {
 		// RToS is active, run threading
 		// Get the mutex so we can use the I2C port
@@ -36,15 +37,23 @@ void FRToSI2C::Mem_Read(uint16_t DevAddress, uint16_t MemAddress,
 
 				I2C1_ClearBusyFlagErratum();
 				xSemaphoreGive(I2CSemaphore);
+				return false;
+			} else {
+				xSemaphoreGive(I2CSemaphore);
+				return true;
 			}
-			xSemaphoreGive(I2CSemaphore);
 #else
 
-			HAL_I2C_Mem_Read(i2c, DevAddress, MemAddress, MemAddSize, pData, Size,
-					5000);
+			if (HAL_I2C_Mem_Read(i2c, DevAddress, MemAddress, MemAddSize, pData, Size,
+					5000)==HAL_OK){
+				xSemaphoreGive(I2CSemaphore);
+				return true;
+			}
 			xSemaphoreGive(I2CSemaphore);
+			return false;
 #endif
 		} else {
+			return false;
 		}
 	}
 
@@ -59,7 +68,7 @@ uint8_t FRToSI2C::I2C_RegisterRead(uint8_t add, uint8_t reg) {
 	return tx_data[0];
 }
 void FRToSI2C::Mem_Write(uint16_t DevAddress, uint16_t MemAddress,
-		uint16_t MemAddSize, uint8_t* pData, uint16_t Size) {
+		uint16_t MemAddSize, uint8_t *pData, uint16_t Size) {
 
 	if (I2CSemaphore == NULL) {
 		// no RToS, run blocking code
@@ -90,7 +99,7 @@ void FRToSI2C::Mem_Write(uint16_t DevAddress, uint16_t MemAddress,
 
 }
 
-void FRToSI2C::Transmit(uint16_t DevAddress, uint8_t* pData, uint16_t Size) {
+void FRToSI2C::Transmit(uint16_t DevAddress, uint8_t *pData, uint16_t Size) {
 	if (I2CSemaphore == NULL) {
 		// no RToS, run blocking code
 		HAL_I2C_Master_Transmit(i2c, DevAddress, pData, Size, 5000);
@@ -119,71 +128,15 @@ void FRToSI2C::Transmit(uint16_t DevAddress, uint8_t* pData, uint16_t Size) {
 
 }
 
-void FRToSI2C::I2C1_ClearBusyFlagErratum() {
-	GPIO_InitTypeDef GPIO_InitStruct;
-	int timeout = 100;
-	int timeout_cnt = 0;
-
-	// 1. Clear PE bit.
-	i2c->Instance->CR1 &= ~(0x0001);
-	/**I2C1 GPIO Configuration
-	 PB6     ------> I2C1_SCL
-	 PB7     ------> I2C1_SDA
-	 */
-	//  2. Configure the SCL and SDA I/Os as General Purpose Output Open-Drain, High level (Write 1 to GPIOx_ODR).
-	GPIO_InitStruct.Mode = GPIO_MODE_OUTPUT_OD;
-	GPIO_InitStruct.Pull = GPIO_PULLUP;
-	GPIO_InitStruct.Speed = GPIO_SPEED_FREQ_LOW;
-
-	GPIO_InitStruct.Pin = SCL_Pin;
-	HAL_GPIO_Init(SCL_GPIO_Port, &GPIO_InitStruct);
-	HAL_GPIO_WritePin(SCL_GPIO_Port, SCL_Pin, GPIO_PIN_SET);
-
-	GPIO_InitStruct.Pin = SDA_Pin;
-	HAL_GPIO_Init(SDA_GPIO_Port, &GPIO_InitStruct);
-	HAL_GPIO_WritePin(SDA_GPIO_Port, SDA_Pin, GPIO_PIN_SET);
-
-	while (GPIO_PIN_SET != HAL_GPIO_ReadPin(SDA_GPIO_Port, SDA_Pin)) {
-		//Move clock to release I2C
-		HAL_GPIO_WritePin(SCL_GPIO_Port, SCL_Pin, GPIO_PIN_RESET);
-		asm("nop");
-		asm("nop");
-		asm("nop");
-		asm("nop");
-		HAL_GPIO_WritePin(SCL_GPIO_Port, SCL_Pin, GPIO_PIN_SET);
-
-		timeout_cnt++;
-		if (timeout_cnt > timeout)
-			return;
+bool FRToSI2C::probe(uint16_t DevAddress) {
+	uint8_t buffer[1];
+	if (Mem_Read(DevAddress, 0, I2C_MEMADD_SIZE_8BIT, buffer, 1)) {
+		//ACK'd
+		return true;
 	}
+	return false;
+}
 
-	// 12. Configure the SCL and SDA I/Os as Alternate function Open-Drain.
-	GPIO_InitStruct.Mode = GPIO_MODE_AF_OD;
-	GPIO_InitStruct.Pull = GPIO_PULLUP;
-	GPIO_InitStruct.Speed = GPIO_SPEED_FREQ_LOW;
-
-	GPIO_InitStruct.Pin = SCL_Pin;
-	HAL_GPIO_Init(SCL_GPIO_Port, &GPIO_InitStruct);
-
-	GPIO_InitStruct.Pin = SDA_Pin;
-	HAL_GPIO_Init(SDA_GPIO_Port, &GPIO_InitStruct);
-
-	HAL_GPIO_WritePin(SCL_GPIO_Port, SCL_Pin, GPIO_PIN_SET);
-	HAL_GPIO_WritePin(SDA_GPIO_Port, SDA_Pin, GPIO_PIN_SET);
-
-	// 13. Set SWRST bit in I2Cx_CR1 register.
-	i2c->Instance->CR1 |= 0x8000;
-
-	asm("nop");
-
-	// 14. Clear SWRST bit in I2Cx_CR1 register.
-	i2c->Instance->CR1 &= ~0x8000;
-
-	asm("nop");
-
-	// 15. Enable the I2C peripheral by setting the PE bit in I2Cx_CR1 register
-	i2c->Instance->CR1 |= 0x0001;
-
-	// Call initialization function.
-	HAL_I2C_Init(i2c);
+void FRToSI2C::I2C1_ClearBusyFlagErratum() {
+	unstick_I2C();
 }
