+++ workspace/TS100/Core/Src/GUIThread.cpp
@@ -19,14 +19,13 @@
 #include "TipThermoModel.h"
 #include "unit.h"
 #include "../../configuration.h"
-
+#include "Buttons.hpp"
 extern uint8_t PCBVersion;
 // File local variables
 extern uint32_t currentTempTargetDegC;
 extern uint8_t accelInit;
 extern uint32_t lastMovementTime;
 extern int16_t idealQCVoltage;
-uint32_t lastButtonTime = 0;
 extern osThreadId GUITaskHandle;
 extern osThreadId MOVTaskHandle;
 extern osThreadId PIDTaskHandle;
@@ -87,113 +86,8 @@ void gui_drawTipTemp(bool symbol) {
 		}
 	}
 }
-ButtonState getButtonState() {
-	/*
-	 * Read in the buttons and then determine if a state change needs to occur
-	 */
 
-	/*
-	 * If the previous state was  00 Then we want to latch the new state if
-	 * different & update time
-	 * If the previous state was !00 Then we want to search if we trigger long
-	 * press (buttons still down), or if release we trigger press
-	 * (downtime>filter)
-	 */
-	static uint8_t previousState = 0;
-	static uint32_t previousStateChange = 0;
-	const uint16_t timeout = 40;
-	uint8_t currentState;
-	currentState = (
-			HAL_GPIO_ReadPin(KEY_A_GPIO_Port, KEY_A_Pin) == GPIO_PIN_RESET ?
-					1 : 0) << 0;
-	currentState |= (
-			HAL_GPIO_ReadPin(KEY_B_GPIO_Port, KEY_B_Pin) == GPIO_PIN_RESET ?
-					1 : 0) << 1;
 
-	if (currentState)
-		lastButtonTime = xTaskGetTickCount();
-	if (currentState == previousState) {
-		if (currentState == 0)
-			return BUTTON_NONE;
-		if ((xTaskGetTickCount() - previousStateChange) > timeout) {
-			// User has been holding the button down
-			// We want to send a buttong is held message
-			if (currentState == 0x01)
-				return BUTTON_F_LONG;
-			else if (currentState == 0x02)
-				return BUTTON_B_LONG;
-			else
-				return BUTTON_NONE; // Both being held case, we dont long hold this
-		} else
-			return BUTTON_NONE;
-	} else {
-		// A change in button state has occurred
-		ButtonState retVal = BUTTON_NONE;
-		if (currentState) {
-			// User has pressed a button down (nothing done on down)
-			if (currentState != previousState) {
-				// There has been a change in the button states
-				// If there is a rising edge on one of the buttons from double press we
-				// want to mask that out As users are having issues with not release
-				// both at once
-				if (previousState == 0x03)
-					currentState = 0x03;
-			}
-		} else {
-			// User has released buttons
-			// If they previously had the buttons down we want to check if they were <
-			// long hold and trigger a press
-			if ((xTaskGetTickCount() - previousStateChange) < timeout) {
-				// The user didn't hold the button for long
-				// So we send button press
-
-				if (previousState == 0x01)
-					retVal = BUTTON_F_SHORT;
-				else if (previousState == 0x02)
-					retVal = BUTTON_B_SHORT;
-				else
-					retVal = BUTTON_BOTH;  // Both being held case
-			}
-		}
-		previousState = currentState;
-		previousStateChange = xTaskGetTickCount();
-		return retVal;
-	}
-	return BUTTON_NONE;
-}
-
-void waitForButtonPress() {
-	// we are just lazy and sleep until user confirms button press
-	// This also eats the button press event!
-	ButtonState buttons = getButtonState();
-	while (buttons) {
-		buttons = getButtonState();
-		GUIDelay();
-	}
-	while (!buttons) {
-		buttons = getButtonState();
-		GUIDelay();
-	}
-}
-
-void waitForButtonPressOrTimeout(uint32_t timeout) {
-	timeout += xTaskGetTickCount();
-	// calculate the exit point
-
-	ButtonState buttons = getButtonState();
-	while (buttons) {
-		buttons = getButtonState();
-		GUIDelay();
-		if (xTaskGetTickCount() > timeout)
-			return;
-	}
-	while (!buttons) {
-		buttons = getButtonState();
-		GUIDelay();
-		if (xTaskGetTickCount() > timeout)
-			return;
-	}
-}
 #ifdef MODEL_TS100
 // returns true if undervoltage has occured
 static bool checkVoltageForExit() {
