+++ workspace/TS100/Core/Src/gui.cpp
@@ -13,7 +13,7 @@
 #include "string.h"
 #include "unit.h"
 #include "../../configuration.h"
-
+#include "Buttons.hpp"
 extern uint32_t lastButtonTime;
 void gui_Menu(const menuitem *menu);
