+++ workspace/TS100/Core/Src/power.cpp
@@ -7,7 +7,7 @@
 
 #include <power.hpp>
 #include <Settings.h>
-#include <hardware.h>
+#include <BSP.h>
 
 const uint16_t powerPWM = 255;
 const uint16_t totalPWM = 255 + 17; //htim2.Init.Period, the full PWM cycle
